refac(web): #5.1 chages name files to route naming
