+++ app/styles/tailwind.css
@@ -1,9 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-.react-aria-TextField {
-  [slot=description] {
-    font-size: 12px;
-  }
-}
